requirements.txt

future
futures

fusepy==3.0.1

simplejson
flatten_json
addict
